move callback to common
