+++ examples/tts/tacotron2.py
@@ -14,7 +14,7 @@
 
 import pytorch_lightning as pl
 
-from nemo.collections.tts.callbacks import LogEpochTimeCallback
+from nemo.collections.common.callbacks import LogEpochTimeCallback
 from nemo.collections.tts.models import Tacotron2
 from nemo.core.config import hydra_runner
 from nemo.utils.exp_manager import exp_manager
