+++ nemo/collections/common/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import nemo.collections.common.callbacks
 from nemo.collections.common import losses, parts, tokenizers
