+++ nemo/collections/common/callbacks/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from nemo.collections.tts.callbacks.callbacks import LogEpochTimeCallback
+from nemo.collections.common.callbacks.callbacks import LogEpochTimeCallback
